requirements.txt
fastapi
uvicorn
gnews
openai
python-dotenv
